+++ examples/separate_compilation/squareAdd.h
@@ -1,3 +1,3 @@
 #pragma once
 
-__device__ __host__ void squareAdd(double& c, double a, double b);
+__device__ __host__ void squareAdd(float& c, float a, float b);
